Move interesting tests over
